Generalize verbose response logging
